chore: view 파일 분리
