+++ src/back-end/web/videos/urls.py
@@ -1,6 +1,7 @@
 """videos URL Configuration"""
 from django.urls import path
-from videos.views import DetailView, HomeView
+from videos.detail_views import DetailView
+from videos.home_view import HomeView
 
 urlpatterns = [
     path("", HomeView.as_view({"get": "list"}), name="home_video_list"),
